+++ lib/cloudshaper.rb
@@ -1,0 +1,6 @@
+require 'cloudshaper/stack'
+require 'cloudshaper/stacks'
+require 'cloudshaper/stack_module'
+require 'cloudshaper/version'
+
+require 'tasks/tasks' if defined? Rake
