+++ lib/cloudshaper/aws/remote_s3.rb
@@ -1,4 +1,4 @@
-module Terraform
+module Cloudshaper
   module Aws
     # Support AWS S3 remote state backend
     module RemoteS3
