+++ lib/cloudshaper/output.rb
@@ -1,0 +1,6 @@
+require 'cloudshaper/stack_element'
+
+module Cloudshaper
+  class Output < StackElement
+  end
+end
