+++ lib/cloudshaper/secrets.rb
