+++ lib/cloudshaper/variable.rb
@@ -1,0 +1,6 @@
+require 'cloudshaper/stack_element'
+
+module Cloudshaper
+  class Variable < StackElement
+  end
+end
