--- lib/terraform_dsl.rb
@@ -1,6 +1,0 @@
-require 'terraform_dsl/stack'
-require 'terraform_dsl/stacks'
-require 'terraform_dsl/stack_module'
-require 'terraform_dsl/version'
-
-require 'tasks/tasks' if defined? Rake
