--- lib/terraform_dsl/output.rb
@@ -1,6 +1,0 @@
-require 'terraform_dsl/stack_element'
-
-module Terraform
-  class Output < StackElement
-  end
-end
