--- lib/terraform_dsl/variable.rb
@@ -1,6 +1,0 @@
-require 'terraform_dsl/stack_element'
-
-module Terraform
-  class Variable < StackElement
-  end
-end
