+++ test/test_helper.rb
@@ -1,4 +1,4 @@
 require 'minitest/autorun'
 require 'minitest/pride'
 
-require 'terraform_dsl'
+require 'cloudshaper'
